+++ src/modules/verticalcoord_mod.f90
@@ -693,12 +693,14 @@ contains
     call rpn_comm_bcast(vco%nlev_depth  , 1, 'MPI_INTEGER', 0, 'GRID', ierr)
     call rpn_comm_bcast(vco%Vcode       , 1, 'MPI_INTEGER', 0, 'GRID', ierr)
     call rpn_comm_bcast(vco%nlev_other, vnl_numvarmaxOther, 'MPI_INTEGER', 0, 'GRID', ierr)
-    if ( mpi_myid > 0 .and. vco%nLev_depth > 0 ) then
-      allocate(vco%ip1_depth(vco%nlev_depth))
-      allocate(vco%depths(vco%nlev_depth))
+    if (vco%nLev_depth > 0) then
+      if ( mpi_myid > 0 ) then
+        allocate(vco%ip1_depth(vco%nlev_depth))
+        allocate(vco%depths(vco%nlev_depth))
+      end if
+      call rpn_comm_bcast(vco%ip1_depth , vco%nlev_depth, 'MPI_INTEGER', 0, 'GRID', ierr)
+      call rpn_comm_bcast(vco%depths    , vco%nlev_depth, 'MPI_REAL8'  , 0, 'GRID', ierr)
     end if
-    call rpn_comm_bcast(vco%ip1_depth , vco%nlev_depth, 'MPI_INTEGER', 0, 'GRID', ierr)
-    call rpn_comm_bcast(vco%depths    , vco%nlev_depth, 'MPI_REAL8'  , 0, 'GRID', ierr)
     if (vco%vgridPresent) then
       if ( mpi_myid == 0 ) then
         vgd_nlev_M = size(vco%ip1_M)